+++ app/Http/Kernel.php
@@ -44,6 +44,7 @@ class Kernel extends HttpKernel
         'throttle' => \Illuminate\Routing\Middleware\ThrottleRequests::class,
         'verified' => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,
         'demo' => \App\Http\Middleware\BlockInDemoMode::class,
+        'manage' => \App\Http\Middleware\ManageAccess::class,
     ];
 
     protected $middlewarePriority = [